Adds BasicObject to virtual ruby object system
Finally realized what seems to be the source of many of the
annoying, weird bugs I've seen when using RSense in
real-world code: when RSense was first written, for Ruby1.8, there was no BasicObject.

BasicObject, in modern Ruby impls, is the lowest object in the
tree, a clean-slate object with only those methods absolutely
necessary for the RubyVM itself.  It is not widely used in
application code to my knowledge, but is widely used in
libraries including much of the std-lib.  Often, it is used
in conjunction with metaprogramming and method_missing
manipulation to enable the magic behind many of the DSL's and
dynamic API's that make Ruby the language it is.

Thus, it made the weird bugs I was seeing an incredible pain to
nail down.

Hopefully, these changes will be enough to make it work,
without having to change a ton of java code.